feat: 🎸 create BoardGameCafe, CafeGameList files

See merge request s07-bigdata-recom-sub2/S07P22B206!16

+++ yogidice/src/main/java/com/specialization/yogidice/domain/repository/BGGDataRepository.java
@@ -4,8 +4,7 @@ import com.specialization.yogidice.domain.entity.BGGData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
-public interface BGGDataRepository extends JpaRepository<BGGData, Integer>{
+public interface BGGDataRepository extends JpaRepository<BGGData, Long> {
     List<BGGData> findBGGDataById(Integer id);
 }
